# DPL: cleanup and speedups using neat optimizations based on vtune insights