--- a/src/dpl/include/dpl/Opendp.h
+++ b/src/dpl/include/dpl/Opendp.h
@@ -112,18 +112,7 @@ struct Master
   bool is_multi_row = false;
 };
 
-struct GridMapKey
-{
-  int grid_index;
-  bool operator<(const GridMapKey& other) const
-  {
-    return grid_index < other.grid_index;
-  }
-  bool operator==(const GridMapKey& other) const
-  {
-    return grid_index == other.grid_index;
-  }
-};
+using GridMapKey = int;
 
 class HybridSiteInfo
 {
@@ -142,6 +131,7 @@ struct Cell
   const char* name() const;
   bool inGroup() const { return group_ != nullptr; }
   int64_t area() const;
+  mutable dbSite* cachedSite_ = nullptr;
 
   dbInst* db_inst_ = nullptr;
   int x_ = 0;  // lower left wrt core DBU
@@ -168,10 +158,16 @@ struct Cell
 
   dbSite* getSite() const
   {
+    if (cachedSite_) {
+      return cachedSite_;
+    }
+
     if (!db_inst_ || !db_inst_->getMaster()) {
       return nullptr;
     }
-    return db_inst_->getMaster()->getSite();
+
+    cachedSite_ = db_inst_->getMaster()->getSite();
+    return cachedSite_;
   }
 };
 
@@ -511,7 +507,7 @@ class Opendp
   int gridX(const Cell* cell, int site_width) const;
   int gridPaddedX(const Cell* cell) const;
   int gridPaddedX(const Cell* cell, int site_width) const;
-  int gridY(int y, int row_height) const;
+  int gridY(int y, const Cell* cell) const;
   int gridY(const Cell* cell) const;
   pair<int, int> gridY(int y, const dbSite::RowPattern& grid_sites) const;
   pair<int, int> gridEndY(int y, const dbSite::RowPattern& grid_sites) const;
@@ -520,7 +516,7 @@ class Opendp
   int gridEndX(int x, int site_width) const;
   int gridEndX(const Cell* cell) const;
   int gridEndX(const Cell* cell, int site_width) const;
-  int gridEndY(int y, int row_height) const;
+  int gridEndY(int y, const Cell* cell) const;
   int gridEndY(const Cell* cell) const;
   void setGridPaddedLoc(Cell* cell, int x, int y, int site_width) const;
   std::pair<int, GridInfo> getRowInfo(const Cell* cell) const;