--- a/src/dpl/src/Place.cpp
+++ b/src/dpl/src/Place.cpp
@@ -892,12 +892,7 @@ PixelPt Opendp::binSearch(int x, const Cell* cell, int bin_x, int bin_y) const
              bin_y);
   int x_end = bin_x + gridPaddedWidth(cell);
   int row_height = getRowHeight(cell);
-  auto grid_mapped_entry = grid_info_map_.find(getGridMapKey(cell));
-  if (grid_mapped_entry == grid_info_map_.end()) {
-    logger_->error(
-        DPL, 14, "Cannot find grid info for row height {}.", row_height);
-  }
-  auto grid_info = grid_mapped_entry->second;
+  auto grid_info = *grid_info_vector_.at(getGridMapKey(cell));
   if (bin_y >= grid_info.getRowCount()) {
     return PixelPt();
   }
@@ -1109,17 +1104,13 @@ Point Opendp::legalPt(const Cell* cell,
   if (row_height == -1) {
     row_height = getRowHeight(cell);
   }
-  auto grid_mapped_entry = grid_info_map_.find(getGridMapKey(cell));
-  if (grid_mapped_entry == grid_info_map_.end()) {
-    logger_->error(
-        DPL, 19, "Cannot find grid info for row height {}.", row_height);
-  }
-  auto grid_info = grid_mapped_entry->second;
+
+  auto grid_info = grid_info_vector_.at(getGridMapKey(cell));
   if (site_width == -1) {
     site_width = getSiteWidth(cell);
   }
   int core_x = min(max(0, pt.getX()),
-                   grid_info.getSiteCount() * site_width - cell->width_);
+                   grid_info->getSiteCount() * site_width - cell->width_);
   // Align with row site.
   int grid_x = divRound(core_x, site_width);
   int legal_x = grid_x * site_width;
@@ -1128,17 +1119,17 @@ Point Opendp::legalPt(const Cell* cell,
     int index(0), height(0);
     int last_row_height = INT_MAX;
     if (cell->isHybridParent()) {
-      last_row_height = grid_info.getRowCount() * row_height - cell->height_;
+      last_row_height = grid_info->getRowCount() * row_height - cell->height_;
     } else {
       auto parent = _hybrid_parent.at(cell->getSite());
-      last_row_height = (grid_info.getRowCount() - 1) * parent->getHeight();
+      last_row_height = (grid_info->getRowCount() - 1) * parent->getHeight();
     }
     std::tie(index, height)
-        = gridY(min(max(0, pt.getY()), last_row_height), grid_info.getSites());
+        = gridY(min(max(0, pt.getY()), last_row_height), grid_info->getSites());
     legal_y = height;
   } else {
     int core_y = min(max(0, pt.getY()),
-                     grid_info.getRowCount() * row_height - cell->height_);
+                     grid_info->getRowCount() * row_height - cell->height_);
     int grid_y = divRound(core_y, row_height);
     legal_y = grid_y * row_height;
   }
@@ -1158,11 +1149,7 @@ Point Opendp::legalGridPt(const Cell* cell,
     row_height = getRowHeight(cell);
   }
   Point legal = legalPt(cell, pt, row_height, site_width);
-  if (cell->isHybrid()) {
-    return Point(gridX(legal.getX(), site_width), gridY(cell));
-  }
-  return Point(gridX(legal.getX(), site_width),
-               gridY(legal.getY(), row_height));
+  return Point(gridX(legal.getX(), site_width), gridY(legal.getY(), cell));
 }
 
 Point Opendp::nearestBlockEdge(const Cell* cell,
@@ -1304,22 +1291,21 @@ Point Opendp::pointOffMacro(const Cell& cell)
   Point init = initialLocation(&cell, false);
   int init_x = init.getX();
   int init_y = init.getY();
-  int row_height = getRowHeight(&cell);
   int site_width = site_width_;
 
   auto grid_info = getGridInfo(&cell);
   Pixel* pixel1 = gridPixel(grid_info.getGridIndex(),
                             gridX(init_x, site_width),
-                            gridY(init_y, row_height));
+                            gridY(init_y, &cell));
   Pixel* pixel2 = gridPixel(grid_info.getGridIndex(),
                             gridX(init_x + cell.width_, site_width),
-                            gridY(init_y, row_height));
+                            gridY(init_y, &cell));
   Pixel* pixel3 = gridPixel(grid_info.getGridIndex(),
                             gridX(init_x, site_width),
-                            gridY(init_y + cell.height_, row_height));
+                            gridY(init_y + cell.height_, &cell));
   Pixel* pixel4 = gridPixel(grid_info.getGridIndex(),
                             gridX(init_x + cell.width_, site_width),
-                            gridY(init_y + cell.height_, row_height));
+                            gridY(init_y + cell.height_, &cell));
 
   Cell* block = nullptr;
   if (pixel1 && pixel1->cell && isBlock(pixel1->cell)) {
@@ -1367,8 +1353,8 @@ void Opendp::legalCellPos(dbInst* db_inst)
 
   int site_width = site_width_;
   // transform to grid Pos for align
-  Point legal_grid_pt = Point(gridX(new_pos.getX(), site_width),
-                              gridY(new_pos.getY(), row_height));
+  Point legal_grid_pt
+      = Point(gridX(new_pos.getX(), site_width), gridY(new_pos.getY(), &cell));
   // Transform position on real position
   int x = (legal_grid_pt.getX() + padLeft(&cell)) * site_width_;
   int y = legal_grid_pt.getY() * row_height;
@@ -1448,7 +1434,7 @@ Point Opendp::legalGridPt(const Cell* cell,
     row_height = getRowHeight(cell);
   }
   Point pt = legalPt(cell, padded, row_height, site_width);
-  return Point(gridX(pt.getX(), site_width), gridY(pt.getY(), row_height));
+  return Point(gridX(pt.getX(), site_width), gridY(pt.getY(), cell));
 }
 
 ////////////////////////////////////////////////////////////////