- replacing grid_info_map.at with grid_info_vector.at reduces runtime from 1 second on aes to 0.22 seconds and removed this: `dpl::GridMapKey::operator< 0.220s `
- caching the getSite function in the cell results in significant gains by removing all of these db calls

```
odb::dbTable<odb::_dbInstHdr>::getPtr 0.280s
odb::_dbObject::getObjectPage   0.240s
odb::dbTable<odb::_dbLib>::getPtr 0.220s
odb::_dbObject::getOwner 0.200s 
```

to only  1 `: odb::_dbObject::getObjectPage 0.164s`